refactor: retry 갱신해서 사용할 수 있도록 수정
