+++ src/main/java/ladder/domain/Line.java
@@ -7,7 +7,7 @@ import static ladder.domain.Direction.STAY;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import ladder.domain.generator.DirectionGenerator;
+import ladder.util.generator.DirectionGenerator;
 
 public class Line {
