+++ src/main/java/ladder/domain/Retry.java
@@ -3,8 +3,13 @@ package ladder.domain;
 public class Retry {
 
     private static final int LIMIT = 0;
+    private static final int DEFAULT_VALUE = 5;
 
     private int value;
+
+    public Retry() {
+        this(DEFAULT_VALUE);
+    }
 
     public Retry(final int value) {
         this.value = value;
@@ -12,11 +17,12 @@ public class Retry {
 
     public void decrease() {
         --value;
-    }
-
-    public void checkCount() {
-        if (value <= LIMIT) {
+        if (value < LIMIT) {
             throw new IllegalStateException("재입력 횟수를 초과하였습니다.");
         }
+    }
+
+    public Retry renew() {
+        return new Retry();
     }
 }
