+++ src/main/java/ladder/util/generator/DirectionGenerator.java
@@ -1,4 +1,4 @@
-package ladder.domain.generator;
+package ladder.util.generator;
 
 import ladder.domain.Direction;
