+++ src/main/java/ladder/util/generator/RandomDirectionGenerator.java
@@ -1,4 +1,4 @@
-package ladder.domain.generator;
+package ladder.util.generator;
 
 import java.util.Random;
 import ladder.domain.Direction;
