+++ src/test/java/ladder/domain/LineTest.java
@@ -7,8 +7,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import java.util.stream.Stream;
-import ladder.domain.generator.DirectionGenerator;
-import ladder.domain.generator.TestDirectionGenerator;
+import ladder.util.generator.DirectionGenerator;
+import ladder.util.generator.TestDirectionGenerator;
 import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
